cleanup

+++ tests/Nivara.Tests/AutoDiff/NullHandlingTests.cs
@@ -1,7 +1,6 @@
-using NUnit.Framework;
-using Nivara;
 using Nivara.Extensions.AutoDiff;
 using Nivara.Extensions.AutoDiff.Operations;
+using NUnit.Framework;
 
 namespace Nivara.Tests.AutoDiff;
 
@@ -19,28 +18,28 @@ public class NullHandlingTests
     public void Add_WithNullValues_PropagatesNullsCorrectly()
     {
         // Arrange: Create columns with null values using Nivara's explicit null semantics
-        var aValues = new int?[] { 1, null, 3, null, 5 };
-        var bValues = new int?[] { 2, 4, null, null, 6 };
-        
-        var aColumn = NivaraColumn<int>.CreateFromNullable(aValues);
-        var bColumn = NivaraColumn<int>.CreateFromNullable(bValues);
-        
-        var a = new GradTensor<int>(aColumn, requiresGrad: true);
-        var b = new GradTensor<int>(bColumn, requiresGrad: true);
+        var aValues = new float?[] { 1.0f, null, 3.0f, null, 5.0f };
+        var bValues = new float?[] { 2.0f, 4.0f, null, null, 6.0f };
+
+        var aColumn = NivaraColumn<float>.CreateFromNullable(aValues);
+        var bColumn = NivaraColumn<float>.CreateFromNullable(bValues);
+
+        var a = new GradTensor<float>(aColumn, requiresGrad: true);
+        var b = new GradTensor<float>(bColumn, requiresGrad: true);
 
         // Act
         var result = GradOperations.Add(a, b);
 
         // Assert: Verify null propagation (null + anything = null)
         Assert.That(result.IsNull(0), Is.False, "Position 0: 1 + 2 should not be null");
-        Assert.That(result[0], Is.EqualTo(3), "Position 0: 1 + 2 = 3");
-        
+        Assert.That(result[0], Is.EqualTo(3.0f), "Position 0: 1 + 2 = 3");
+
         Assert.That(result.IsNull(1), Is.True, "Position 1: null + 4 should be null");
         Assert.That(result.IsNull(2), Is.True, "Position 2: 3 + null should be null");
         Assert.That(result.IsNull(3), Is.True, "Position 3: null + null should be null");
-        
+
         Assert.That(result.IsNull(4), Is.False, "Position 4: 5 + 6 should not be null");
-        Assert.That(result[4], Is.EqualTo(11), "Position 4: 5 + 6 = 11");
+        Assert.That(result[4], Is.EqualTo(11.0f), "Position 4: 5 + 6 = 11");
     }
 
     [Test]
@@ -49,10 +48,10 @@ public class NullHandlingTests
         // Arrange
         var aValues = new float?[] { 2.0f, null, 4.0f, null };
         var bValues = new float?[] { 3.0f, 5.0f, null, null };
-        
+
         var aColumn = NivaraColumn<float>.CreateFromNullable(aValues);
         var bColumn = NivaraColumn<float>.CreateFromNullable(bValues);
-        
+
         var a = new GradTensor<float>(aColumn, requiresGrad: true);
         var b = new GradTensor<float>(bColumn, requiresGrad: true);
 
@@ -62,7 +61,7 @@ public class NullHandlingTests
         // Assert: Verify null propagation (null * anything = null)
         Assert.That(result.IsNull(0), Is.False, "Position 0: 2 * 3 should not be null");
         Assert.That(result[0], Is.EqualTo(6.0f), "Position 0: 2 * 3 = 6");
-        
+
         Assert.That(result.IsNull(1), Is.True, "Position 1: null * 5 should be null");
         Assert.That(result.IsNull(2), Is.True, "Position 2: 4 * null should be null");
         Assert.That(result.IsNull(3), Is.True, "Position 3: null * null should be null");
@@ -74,10 +73,10 @@ public class NullHandlingTests
         // Arrange
         var aValues = new double?[] { 10.0, null, 6.0 };
         var bValues = new double?[] { 3.0, 2.0, null };
-        
+
         var aColumn = NivaraColumn<double>.CreateFromNullable(aValues);
         var bColumn = NivaraColumn<double>.CreateFromNullable(bValues);
-        
+
         var a = new GradTensor<double>(aColumn, requiresGrad: true);
         var b = new GradTensor<double>(bColumn, requiresGrad: true);
 
@@ -87,7 +86,7 @@ public class NullHandlingTests
         // Assert: Verify null propagation (null - anything = null)
         Assert.That(result.IsNull(0), Is.False, "Position 0: 10 - 3 should not be null");
         Assert.That(result[0], Is.EqualTo(7.0), "Position 0: 10 - 3 = 7");
-        
+
         Assert.That(result.IsNull(1), Is.True, "Position 1: null - 2 should be null");
         Assert.That(result.IsNull(2), Is.True, "Position 2: 6 - null should be null");
     }
@@ -98,10 +97,10 @@ public class NullHandlingTests
         // Arrange
         var aValues = new float?[] { 12.0f, null, 18.0f };
         var bValues = new float?[] { 3.0f, 5.0f, null };
-        
+
         var aColumn = NivaraColumn<float>.CreateFromNullable(aValues);
         var bColumn = NivaraColumn<float>.CreateFromNullable(bValues);
-        
+
         var a = new GradTensor<float>(aColumn, requiresGrad: true);
         var b = new GradTensor<float>(bColumn, requiresGrad: true);
 
@@ -111,7 +110,7 @@ public class NullHandlingTests
         // Assert: Verify null propagation (null / anything = null)
         Assert.That(result.IsNull(0), Is.False, "Position 0: 12 / 3 should not be null");
         Assert.That(result[0], Is.EqualTo(4.0f), "Position 0: 12 / 3 = 4");
-        
+
         Assert.That(result.IsNull(1), Is.True, "Position 1: null / 5 should be null");
         Assert.That(result.IsNull(2), Is.True, "Position 2: 18 / null should be null");
     }
@@ -134,14 +133,14 @@ public class NullHandlingTests
         // Assert: Verify null propagation
         Assert.That(result.IsNull(0), Is.False, "Position 0: relu(-2) should not be null");
         Assert.That(result[0], Is.EqualTo(0.0f), "Position 0: relu(-2) = 0");
-        
+
         Assert.That(result.IsNull(1), Is.True, "Position 1: relu(null) should be null");
-        
+
         Assert.That(result.IsNull(2), Is.False, "Position 2: relu(0) should not be null");
         Assert.That(result[2], Is.EqualTo(0.0f), "Position 2: relu(0) = 0");
-        
+
         Assert.That(result.IsNull(3), Is.True, "Position 3: relu(null) should be null");
-        
+
         Assert.That(result.IsNull(4), Is.False, "Position 4: relu(2) should not be null");
         Assert.That(result[4], Is.EqualTo(2.0f), "Position 4: relu(2) = 2");
     }
@@ -160,12 +159,12 @@ public class NullHandlingTests
         // Assert: Verify null propagation
         Assert.That(result.IsNull(0), Is.False, "Position 0: sigmoid(0) should not be null");
         Assert.That(result[0], Is.EqualTo(0.5f).Within(0.0001f), "Position 0: sigmoid(0) = 0.5");
-        
+
         Assert.That(result.IsNull(1), Is.True, "Position 1: sigmoid(null) should be null");
-        
+
         Assert.That(result.IsNull(2), Is.False, "Position 2: sigmoid(1) should not be null");
         Assert.That(result[2], Is.EqualTo(0.7311f).Within(0.001f), "Position 2: sigmoid(1) ≈ 0.7311");
-        
+
         Assert.That(result.IsNull(3), Is.True, "Position 3: sigmoid(null) should be null");
     }
 
@@ -183,9 +182,9 @@ public class NullHandlingTests
         // Assert: Verify null propagation
         Assert.That(result.IsNull(0), Is.False, "Position 0: tanh(0) should not be null");
         Assert.That(result[0], Is.EqualTo(0.0).Within(0.0001), "Position 0: tanh(0) = 0");
-        
+
         Assert.That(result.IsNull(1), Is.True, "Position 1: tanh(null) should be null");
-        
+
         Assert.That(result.IsNull(2), Is.False, "Position 2: tanh(1) should not be null");
         Assert.That(result[2], Is.EqualTo(0.7616).Within(0.001), "Position 2: tanh(1) ≈ 0.7616");
     }
@@ -200,35 +199,35 @@ public class NullHandlingTests
         // Arrange
         var aValues = new float?[] { 1.0f, null, 3.0f };
         var bValues = new float?[] { 2.0f, 4.0f, null };
-        
+
         var aColumn = NivaraColumn<float>.CreateFromNullable(aValues);
         var bColumn = NivaraColumn<float>.CreateFromNullable(bValues);
-        
+
         var a = new GradTensor<float>(aColumn, requiresGrad: true);
         var b = new GradTensor<float>(bColumn, requiresGrad: true);
 
         // Act: Forward pass
         var result = GradOperations.Add(a, b);
-        
+
         // Create gradient output (all ones for non-null positions)
         var gradValues = new float?[] { 1.0f, 1.0f, 1.0f };
         var gradColumn = NivaraColumn<float>.CreateFromNullable(gradValues);
         var grad = new GradTensor<float>(gradColumn, requiresGrad: false);
-        
+
         result.Backward(grad);
 
         // Assert: Gradients should be computed for non-null positions
         Assert.That(a.Grad, Is.Not.Null, "Gradient for a should exist");
         Assert.That(b.Grad, Is.Not.Null, "Gradient for b should exist");
-        
+
         // Position 0: both non-null, gradient should flow
         Assert.That(a.Grad![0], Is.EqualTo(1.0f), "Gradient for a[0] should be 1");
         Assert.That(b.Grad![0], Is.EqualTo(1.0f), "Gradient for b[0] should be 1");
-        
+
         // Position 1: a is null, gradient should still accumulate
         Assert.That(a.Grad[1], Is.EqualTo(1.0f), "Gradient for a[1] should be 1");
         Assert.That(b.Grad[1], Is.EqualTo(1.0f), "Gradient for b[1] should be 1");
-        
+
         // Position 2: b is null, gradient should still accumulate
         Assert.That(a.Grad[2], Is.EqualTo(1.0f), "Gradient for a[2] should be 1");
         Assert.That(b.Grad[2], Is.EqualTo(1.0f), "Gradient for b[2] should be 1");
@@ -244,26 +243,26 @@ public class NullHandlingTests
 
         // Act: Forward pass
         var result = GradOperations.Relu(tensor);
-        
+
         // Create gradient output (all ones)
         var gradValues = new float?[] { 1.0f, 1.0f, 1.0f, 1.0f };
         var gradColumn = NivaraColumn<float>.CreateFromNullable(gradValues);
         var grad = new GradTensor<float>(gradColumn, requiresGrad: false);
-        
+
         result.Backward(grad);
 
         // Assert: Gradients should be computed correctly
         Assert.That(tensor.Grad, Is.Not.Null, "Gradient should exist");
-        
+
         // Position 0: input -1 <= 0, gradient should be 0
         Assert.That(tensor.Grad![0], Is.EqualTo(0.0f), "Gradient for negative input should be 0");
-        
+
         // Position 1: input is null, gradient should be 0 (null handling)
         Assert.That(tensor.Grad[1], Is.EqualTo(0.0f), "Gradient for null input should be 0");
-        
+
         // Position 2: input 1 > 0, gradient should be 1
         Assert.That(tensor.Grad[2], Is.EqualTo(1.0f), "Gradient for positive input should be 1");
-        
+
         // Position 3: input 2 > 0, gradient should be 1
         Assert.That(tensor.Grad[3], Is.EqualTo(1.0f), "Gradient for positive input should be 1");
     }
@@ -276,16 +275,16 @@ public class NullHandlingTests
     public void GradTensor_PreservesNullMaskFromNivaraColumn()
     {
         // Arrange
-        var values = new int?[] { 1, null, 3, null, 5 };
-        var column = NivaraColumn<int>.CreateFromNullable(values);
+        var values = new float?[] { 1.0f, null, 3.0f, null, 5.0f };
+        var column = NivaraColumn<float>.CreateFromNullable(values);
 
         // Act
-        var tensor = new GradTensor<int>(column, requiresGrad: true);
+        var tensor = new GradTensor<float>(column, requiresGrad: true);
 
         // Assert: Verify null mask is preserved
         Assert.That(tensor.HasNulls, Is.True, "Tensor should indicate it has nulls");
         Assert.That(tensor.Length, Is.EqualTo(5), "Tensor should have correct length");
-        
+
         for (int i = 0; i < values.Length; i++)
         {
             bool expectedIsNull = values[i] == null;
@@ -300,10 +299,10 @@ public class NullHandlingTests
         // Arrange: Create tensors with explicit null masks
         var aValues = new float?[] { 1.0f, null, 3.0f };
         var bValues = new float?[] { 2.0f, 4.0f, null };
-        
+
         var aColumn = NivaraColumn<float>.CreateFromNullable(aValues);
         var bColumn = NivaraColumn<float>.CreateFromNullable(bValues);
-        
+
         var a = new GradTensor<float>(aColumn, requiresGrad: true);
         var b = new GradTensor<float>(bColumn, requiresGrad: true);
 
@@ -316,12 +315,12 @@ public class NullHandlingTests
         Assert.That(add.HasNulls, Is.True, "Addition result should have nulls");
         Assert.That(add.IsNull(1), Is.True, "Addition: null + 4 should be null");
         Assert.That(add.IsNull(2), Is.True, "Addition: 3 + null should be null");
-        
+
         // Multiplication result
         Assert.That(mul.HasNulls, Is.True, "Multiplication result should have nulls");
         Assert.That(mul.IsNull(1), Is.True, "Multiplication: null * 4 should be null");
         Assert.That(mul.IsNull(2), Is.True, "Multiplication: 3 * null should be null");
-        
+
         // Verify no sentinel values are used (explicit null mask approach)
         // This is implicit in Nivara's design - nulls are tracked via null masks, not special values
     }
@@ -374,10 +373,10 @@ public class NullHandlingTests
         // Arrange: Test a complex expression with nulls
         var xValues = new float?[] { 1.0f, null, 3.0f };
         var wValues = new float?[] { 0.5f, 0.5f, null };
-        
+
         var xColumn = NivaraColumn<float>.CreateFromNullable(xValues);
         var wColumn = NivaraColumn<float>.CreateFromNullable(wValues);
-        
+
         var x = new GradTensor<float>(xColumn, requiresGrad: false);
         var w = new GradTensor<float>(wColumn, requiresGrad: true);
 
@@ -388,14 +387,14 @@ public class NullHandlingTests
         // Assert: Verify null propagation through operations
         Assert.That(mul.IsNull(0), Is.False, "Position 0: 1 * 0.5 should not be null");
         Assert.That(mul[0], Is.EqualTo(0.5f), "Position 0: 1 * 0.5 = 0.5");
-        
+
         Assert.That(mul.IsNull(1), Is.True, "Position 1: null * 0.5 should be null");
         Assert.That(mul.IsNull(2), Is.True, "Position 2: 3 * null should be null");
-        
+
         // ReLU should preserve nulls
         Assert.That(relu.IsNull(0), Is.False, "Position 0: relu(0.5) should not be null");
         Assert.That(relu[0], Is.EqualTo(0.5f), "Position 0: relu(0.5) = 0.5");
-        
+
         Assert.That(relu.IsNull(1), Is.True, "Position 1: relu(null) should be null");
         Assert.That(relu.IsNull(2), Is.True, "Position 2: relu(null) should be null");
     }
@@ -405,7 +404,7 @@ public class NullHandlingTests
     {
         // This test verifies that Nivara's explicit null semantics are preserved
         // and no sentinel values (like NaN, -1, etc.) are used to represent nulls
-        
+
         // Arrange
         var values = new float?[] { 0.0f, null, float.NaN, null, -1.0f };
         var column = NivaraColumn<float>.CreateFromNullable(values);
@@ -414,15 +413,15 @@ public class NullHandlingTests
         // Assert: Verify explicit null tracking
         Assert.That(tensor.IsNull(0), Is.False, "Position 0: 0.0 should not be null");
         Assert.That(tensor[0], Is.EqualTo(0.0f), "Position 0: value should be 0.0");
-        
+
         Assert.That(tensor.IsNull(1), Is.True, "Position 1: should be null (explicit)");
-        
+
         // NaN is a valid value, not a null sentinel
         Assert.That(tensor.IsNull(2), Is.False, "Position 2: NaN should not be treated as null");
         Assert.That(float.IsNaN(tensor[2]), Is.True, "Position 2: value should be NaN");
-        
+
         Assert.That(tensor.IsNull(3), Is.True, "Position 3: should be null (explicit)");
-        
+
         Assert.That(tensor.IsNull(4), Is.False, "Position 4: -1.0 should not be null");
         Assert.That(tensor[4], Is.EqualTo(-1.0f), "Position 4: value should be -1.0");
     }
